requirements.txt
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
Eel==0.14.0
Flask==2.0.1
future==0.18.2
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.1.0
idna==2.10
importlib-metadata==4.5.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pycparser==2.20
pyparsing==2.4.7
requests==2.25.1
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==2.0.1
whichcraft==0.6.1
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.4.0
alembic==1.6.5
Babel==2.9.1
bcrypt==3.2.0
blinker==1.4
Brotli==1.0.9
cffi==1.14.5
click==8.0.1
dash==1.20.0
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
Flask-SQLAlchemy==2.5.1
future==0.18.2
greenlet==1.1.0
itsdangerous==2.0.1
Mako==1.1.4
MarkupSafe==2.0.1
passlib==1.7.4
plotly==4.14.3
py-dotenv==0.1
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
pytz==2021.1
retrying==1.3.3
six==1.16.0
speaklater==1.3
SQLAlchemy==1.4.17

